Уносит константы ближе к парсерам. Добавляет интерфейс IRenderMode.
