+++ src/main/java/ru/naumen/sd40/log/parser/Constants.java
@@ -1,0 +1,7 @@
+package ru.naumen.sd40.log.parser;
+
+public class Constants
+{
+    public static final String MEASUREMENT_NAME = "perf";
+    public static final String TIME = "time";
+}
