+++ src/main/java/ru/naumen/sd40/log/parser/Gc/GcMode.java
@@ -3,6 +3,7 @@ package ru.naumen.sd40.log.parser.Gc;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import ru.naumen.perfhouse.influx.InfluxDAO;
+import ru.naumen.sd40.log.parser.DataType;
 import ru.naumen.perfhouse.writers.GcInfluxWriter;
 import ru.naumen.perfhouse.writers.IDatabaseWriter;
 import ru.naumen.sd40.log.parser.IDataParser;
@@ -14,6 +15,7 @@ public class GcMode implements IParsingMode {
     private GcDataParser dataParser;
     private GcTimeParserFactory timeParserFactory;
     private GcDataSetFactory dataSetFactory;
+    private DataType dataType = new DataType(GcConstants.getProps());
 
     @Autowired
     GcMode(GcDataParser dataParser, GcTimeParserFactory timeParserFactory, GcDataSetFactory dataSetFactory) {
@@ -40,5 +42,20 @@ public class GcMode implements IParsingMode {
     @Override
     public GcDataSetFactory getDataSetFactory() {
         return dataSetFactory;
+    }
+
+    @Override
+    public String getView() {
+        return "history_gc";
+    }
+
+    @Override
+    public DataType getDataType() {
+        return dataType;
+    }
+
+    @Override
+    public String getBeautifulParserName() {
+        return "Garbage Collection";
     }
 }
