+++ src/main/java/ru/naumen/sd40/log/parser/IRenderMode.java
@@ -1,0 +1,7 @@
+package ru.naumen.sd40.log.parser;
+
+public interface IRenderMode {
+    String getView();
+    String getBeautifulParserName();
+    DataType getDataType();
+}
